fix: merge club registration and member registraion modules to registration module
