+++ packages/api/src/feature/registration/model/member.registration.model.ts
