+++ packages/api/src/feature/registration/registration.module.ts
@@ -1,15 +1,32 @@
-import { Module } from "@nestjs/common";
-import { DrizzleModule } from "src/drizzle/drizzle.module";
+import { forwardRef, Module } from "@nestjs/common";
 
-import { ClubRegistrationModule } from "./club-registration/club-registration.module";
+import { DrizzleModule } from "@sparcs-clubs/api/drizzle/drizzle.module";
+import ClubModule from "@sparcs-clubs/api/feature/club/club.module";
+import DivisionModule from "@sparcs-clubs/api/feature/division/division.module";
+import FileModule from "@sparcs-clubs/api/feature/file/file.module";
+import UserModule from "@sparcs-clubs/api/feature/user/user.module";
+
 import { RegistrationController } from "./controller/registration.controller";
-import { MemberRegistrationModule } from "./member-registration/member-registration.module";
-import { RegistrationRepository } from "./repository/registration.repository";
+import { ClubRegistrationRepository } from "./repository/club-registration.repository";
+import { MemberRegistrationRepository } from "./repository/member-registration.repository";
+import { RegistrationPublicService } from "./service/registration.public.service";
 import { RegistrationService } from "./service/registration.service";
 
 @Module({
-  imports: [DrizzleModule, MemberRegistrationModule, ClubRegistrationModule],
+  imports: [
+    DrizzleModule,
+    forwardRef(() => ClubModule),
+    DivisionModule,
+    FileModule,
+    UserModule,
+  ],
   controllers: [RegistrationController],
-  providers: [RegistrationRepository, RegistrationService],
+  providers: [
+    RegistrationService,
+    RegistrationPublicService,
+    ClubRegistrationRepository,
+    MemberRegistrationRepository,
+  ],
+  exports: [RegistrationPublicService],
 })
 export class RegistrationModule {}
